+++ src/main/java/com/apis/fintrack/Service/TransactionService.java
@@ -5,6 +5,7 @@ import com.apis.fintrack.Entity.TransactionEntity;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public interface TransactionService {
@@ -13,7 +14,7 @@ public interface TransactionService {
     Page<TransactionEntity> findAll(Pageable pageable);
     Page<TransactionEntity> findAllByCategory(CategoryEnum category, Pageable pageable);
     Page<TransactionEntity> findAllBetween(LocalDate startDate, LocalDate endDate, Pageable pageable);
-    Page<TransactionEntity> findAllBetweenAmounts(Double firstAmount, Double secondAmount, Pageable pageable);
+    Page<TransactionEntity> findAllBetweenAmounts(BigDecimal firstAmount, BigDecimal secondAmount, Pageable pageable);
     Page<TransactionEntity> findAllOrderByAmount(Pageable pageable);
     Page<TransactionEntity> findAllOrderByTransactionDate(Pageable pageable);
     Page<TransactionEntity> findAllIfIncome(Boolean isIncome, Pageable pageable);
